+++ JobRecruiting/JobRecruiting/src/main/java/com/recruit/jobrecruiting/interviews/ejb/InterviewBean.java
@@ -7,7 +7,7 @@ package com.recruit.jobrecruiting.interviews.ejb;
 
 import com.recruit.jobrecruiting.common.InterviewDetails;
 import com.recruit.jobrecruiting.common.UserDetails;
-import com.recruit.jobrecruiting.ejb.JobPostBean;
+import com.recruit.jobrecruiting.jobPost.ejb.JobPostBean;
 import com.recruit.jobrecruiting.entity.Interview;
 import com.recruit.jobrecruiting.entity.InterviewStatus;
 import com.recruit.jobrecruiting.entity.JobPost;
