+++ JobRecruiting/JobRecruiting/src/main/java/com/recruit/jobrecruiting/interviews/ejb/ViewCandidatesBean.java
@@ -6,7 +6,7 @@
 package com.recruit.jobrecruiting.interviews.ejb;
 
 import com.recruit.jobrecruiting.common.InterviewDetails;
-import com.recruit.jobrecruiting.ejb.JobPostBean;
+import com.recruit.jobrecruiting.jobPost.ejb.JobPostBean;
 import com.recruit.jobrecruiting.entity.Interview;
 import com.recruit.jobrecruiting.entity.InterviewStatus;
 import com.recruit.jobrecruiting.user.ejb.UserBean;
