+++ JobRecruiting/JobRecruiting/src/main/java/com/recruit/jobrecruiting/jobPost/servlet/JobPost.java
@@ -4,12 +4,13 @@
  * and open the template in the editor.
  */
 
-package com.recruit.jobrecruiting.servlet.jobpost;
+package com.recruit.jobrecruiting.jobpost.servlet;
 
-import com.recruit.jobrecruiting.ejb.JobPostBean;
-import com.recruit.jobrecruiting.entity.Status;
-import com.recruit.jobrecruiting.mail.EmailBean;
-import com.recruit.jobrecruiting.util.Util;
+import com.recruit.jobrecruiting.common.JobPostDetails;
+import com.recruit.jobrecruiting.jobPost.ejb.JobPostBean;
+import com.recruit.jobrecruiting.entity.User;
+import com.recruit.jobrecruiting.interviews.ejb.InterviewBean;
+import com.recruit.jobrecruiting.user.ejb.UserBean;
 import java.io.IOException;
 import javax.inject.Inject;
 import javax.servlet.ServletException;
@@ -24,15 +25,18 @@ import javax.servlet.http.HttpServletResponse;
  *
  * @author DENISA
  */
-@ServletSecurity(value = @HttpConstraint(rolesAllowed = {"RecruiterRole"}))
-@WebServlet(name="CopyJobPost", urlPatterns={"/JobPost/Copy"})
-public class CopyJobPost extends HttpServlet {
+@ServletSecurity(value = @HttpConstraint(rolesAllowed = {"CandidateRole"}))
+@WebServlet(name = "JobPost", urlPatterns = {"/JobPost"})
+public class JobPost extends HttpServlet {
 
     @Inject
     private JobPostBean jobPostBean;
 
     @Inject
-    private EmailBean emailBean;
+    private InterviewBean interviewBean;
+
+    @Inject
+    private UserBean userBean;
 
     /** 
      * Handles the HTTP <code>GET</code> method.
@@ -43,8 +47,16 @@ public class CopyJobPost extends HttpServlet {
      */
     @Override
     protected void doGet(HttpServletRequest request, HttpServletResponse response)
-    throws ServletException, IOException {
-        
+            throws ServletException, IOException {
+
+        int id = Integer.parseInt(request.getParameter("id"));
+        JobPostDetails jobPost = jobPostBean.getJobPost(id);
+        request.setAttribute("jobPost", jobPost);
+
+        User user = userBean.getUserByUsername(request.getRemoteUser());
+        request.setAttribute("jobPostsAppliedToIds", interviewBean.getAllJobPostsAsCandidate(user.getId()));
+
+        request.getRequestDispatcher("/WEB-INF/pages/jobpost/jobpost.jsp").forward(request, response);
     } 
 
     /** 
@@ -56,23 +68,11 @@ public class CopyJobPost extends HttpServlet {
      */
     @Override
     protected void doPost(HttpServletRequest request, HttpServletResponse response)
-    throws ServletException, IOException {
+            throws ServletException, IOException {
+
         int id = Integer.parseInt(request.getParameter("id"));
-
-        Status status = Status.WAITING_FOR_APPROVAL;
-        if (request.isUserInRole("GeneralDirectorRole")) {
-            status = Status.INACTIVE;
-        }
-
-        String username = request.getRemoteUser();
-        int jobpost_id = jobPostBean.copyJobPost(id, status, username).getId();
-
-        if (!request.isUserInRole("GeneralDirectorRole")) {
-            sendEmail(request, jobpost_id);
-
-        }
-        response.sendRedirect(request.getContextPath() + "/JobPost/Edit?id=" + jobpost_id);
-
+        jobPostBean.deleteJobPost(id);
+        response.sendRedirect(request.getContextPath() + "/JobPosts");
     }
 
     /** 
@@ -82,13 +82,6 @@ public class CopyJobPost extends HttpServlet {
     @Override
     public String getServletInfo() {
         return "Short description";
-    }
-
-    protected void sendEmail(HttpServletRequest request, int jobpost_id) {
-        new Thread(() -> {
-            String url = Util.getBaseUrl(request) + "/JobPost?id=" + jobpost_id;
-            emailBean.sendEmail("denisa.halmaghi@ulbsibiu.ro", "New jobpost created", url);
-        }).start();
     }
 
 }
